Move Footer to views
